Rating added to the Product Model

+++ lib/models/product.dart
@@ -32,7 +32,7 @@ class Product {
 
 class Rating {
   final double? rate;
-  final double? count;
+  final int? count;
 
   Rating({this.rate, this.count});
